requirements.txt
readability-lxml
requests
cssselect
cpsk==0.1.1
pygeocoder
brutal-ng==0.3.10
imhdsk
workalendar
pynameday
feedparser
pyshorteners
hladny_matfyzak
